+++ test/canary.js
@@ -1,7 +1,7 @@
-const assert = require("assert");
+const assert = require('assert');
 
-describe("eslint-config-apostrophe:canary", function() {
-  it("should pass a canary test", function() {
+describe('eslint-config-apostrophe:canary', function() {
+  it('should pass a canary test', function() {
     // eslint-disable-next-line
     assert.equal(true, true);
   });