+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Ignores inside CSS and preprocessor functions for the `declaration-property-unit-allowed-list` rule. Those functions do not represent the ultimate property unit after compiling.
+- Adds eslint to test script and uses Apostrophe eslint config package.
 
 ## [1.0.1] - 2021-03-24